[#5] ✨ issue api 개선
