+++ libs/types/common.ts
@@ -1,0 +1,4 @@
+export interface OctokitRestCommonParamsType {
+    owner: string
+    repo: string
+}
