+++ libs/types/index.ts
@@ -1,11 +1,3 @@
-export interface OctokitRestCommonParamsType {
-    owner: string
-    repo: string
-}
+export * from './common'
 
-export interface CreateIssueParams {
-    title: string
-    body?: string
-    assignee?: string
-    labels?: string[]
-}
+export * from './issue'
